create subcommand and move some commands under it
