+++ cmd/config/config.go
@@ -1,14 +1,11 @@
-package cmd
+package config
 
 import (
 	"github.com/spf13/cobra"
 )
 
-func init() {
-	RootCmd.AddCommand(configCmd)
-}
-
-var configCmd = &cobra.Command{
+// ConfigCmd represents the config subcommand.
+var ConfigCmd = &cobra.Command{
 	Use:   "config",
 	Short: "Manage the forge.yaml configuration",
 	Long:  `Commands for initializing and formatting the forge.yaml configuration file.`,
