+++ cmd/config/fmt.go
@@ -1,25 +1,24 @@
-package cmd
+package config
 
 import (
 	"fmt"
 
-	"github.com/spf13/cobra"
-
 	"github.com/gomicro/forge/confile"
+	"github.com/spf13/cobra"
 )
 
 func init() {
-	configCmd.AddCommand(confmtCmd)
+	ConfigCmd.AddCommand(fmtCmd)
 }
 
-var confmtCmd = &cobra.Command{
+var fmtCmd = &cobra.Command{
 	Use:   "fmt",
 	Short: "Format the forge.yaml config file",
 	Long:  `Rewrite forge.yaml in the current directory in canonical format.`,
-	RunE:  confmtFunc,
+	RunE:  fmtFunc,
 }
 
-func confmtFunc(cmd *cobra.Command, args []string) error {
+func fmtFunc(cmd *cobra.Command, args []string) error {
 	conf, err := confile.ParseFromFile()
 	if err != nil {
 		return fmt.Errorf("parsing config file: %w", err)
